+++ loggers/winston/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## v0.3.0
 
 - Serialize "log.level" as a top-level dotted field per
-  <https://github.com/elastic/ecs-logging/pull/33> -
+  https://github.com/elastic/ecs-logging/pull/33.
   [#23](https://github.com/elastic/ecs-logging-js/pull/23)
 
 ## v0.2.0