+++ loggers/winston/CHANGELOG.md
@@ -6,7 +6,7 @@
   done by default. One must use the new `convertReqRes: true` formatter option.
   As well, only the meta keys `req` and `res` will be handled. Before this
   change the meta keys `req`, `res`, `request`, and `response` would all be
-  handled.
+  handled. (Issue #32)
 
   Before (no longer works):